backend/requirements.txt
# SafeGuide Backend Dependencies
# Cleaned for Railway deployment — removed unused heavy packages

# ── Core Flask ────────────────────────────────────────────────────────────
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3

# ── Database ──────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9      # PostgreSQL driver (Railway Postgres)

# ── AI ───────────────────────────────────────────────────────────────────
groq==0.9.0                  # Groq API client (replaces anthropic)

# ── Auth & Security ───────────────────────────────────────────────────────
bcrypt==4.0.1

# ── Utilities ────────────────────────────────────────────────────────────
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0             # Production WSGI server for Railway